refactor(sandbox): share buildConfigPayload across runners
Extract the duplicated buildConfigPayload and deriveRepoLabel helpers from
agent-sandbox, docker, host, and freestyle runners into
runner/shared/build-config-payload.ts so the daemon /config payload shape
stays in lockstep across all runners. Renames host's local devPort param
to desiredPort to match the canonical TenantConfig field name.
