+++ packages/sandbox/server/runner/shared/index.ts
@@ -2,3 +2,7 @@ export { Inflight } from "./inflight";
 export { withSandboxLock } from "./lock";
 export { computeHandle, hashSandboxId } from "./handle";
 export { applyPreviewPattern } from "./preview-url";
+export {
+  buildConfigPayload,
+  deriveRepoLabel,
+} from "./build-config-payload";
